Refactored nodes.rs
